# simd8x64::compress optimisation for Apple M1