This is an optimisation of the arm64 implementation of `simd8x64::compress` for the Apple M1, which increases `fast_minify` benchmark speed by ~39%. The idea is to do twice as many stores, but significantly fewer loads and SIMD operations. I also reduce the number of population counts (as they are four instructions on ARM64, with 15c latency on Apple M1 Firestorm). This PR has not been tested on other CPUs.

The second commit could be removed, as it only accounts for a 3-3.5% speedup, and required changes to non-ARM64-specific code. Theses changes shouldn't change behaviour, performance or code generation, but I'm not set up to test properly.

This PR:
```
------------------------------------------------------------------------------------------------
Benchmark                                      Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------
fast_minify_twitter/repeats:10_mean        68231 ns        68231 ns           10 Gigabytes=9.25554G/s docs=14.6561k/s
fast_minify_twitter/repeats:10_median      68218 ns        68218 ns           10 Gigabytes=9.25733G/s docs=14.6589k/s
fast_minify_twitter/repeats:10_stddev       35.6 ns         35.5 ns           10 Gigabytes=4.81706M/s docs=7.62778/s
fast_minify_twitter/repeats:10_max         68299 ns        68299 ns           10 Gigabytes=9.25965G/s docs=14.6626k/s
fast_minify_gsoc/repeats:10_mean          367660 ns       367660 ns           10 Gigabytes=9.05138G/s docs=2.7199k/s
fast_minify_gsoc/repeats:10_median        367639 ns       367639 ns           10 Gigabytes=9.05189G/s docs=2.72006k/s
fast_minify_gsoc/repeats:10_stddev          72.2 ns         71.8 ns           10 Gigabytes=1.76873M/s docs=0.531496/s
fast_minify_gsoc/repeats:10_max           367746 ns       367747 ns           10 Gigabytes=9.05423G/s docs=2.72076k/s
```

Main branch:
```
------------------------------------------------------------------------------------------------
Benchmark                                      Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------
fast_minify_twitter/repeats:10_mean        95862 ns        95861 ns           10 Gigabytes=6.58779G/s docs=10.4317k/s
fast_minify_twitter/repeats:10_median      95858 ns        95857 ns           10 Gigabytes=6.58806G/s docs=10.4322k/s
fast_minify_twitter/repeats:10_stddev       45.5 ns         45.6 ns           10 Gigabytes=3.12853M/s docs=4.954/s
fast_minify_twitter/repeats:10_max         95970 ns        95970 ns           10 Gigabytes=6.59135G/s docs=10.4374k/s
fast_minify_gsoc/repeats:10_mean          514316 ns       514314 ns           10 Gigabytes=6.47042G/s docs=1.94434k/s
fast_minify_gsoc/repeats:10_median        514275 ns       514275 ns           10 Gigabytes=6.47092G/s docs=1.94449k/s
fast_minify_gsoc/repeats:10_stddev           173 ns          172 ns           10 Gigabytes=2.16759M/s docs=0.651353/s
fast_minify_gsoc/repeats:10_max           514727 ns       514727 ns           10 Gigabytes=6.47256G/s docs=1.94498k/s
```

And using [docker_programming_station](https://github.com/lemire/docker_programming_station):
 
This PR (GCC 10):
```
------------------------------------------------------------------------------------------------
Benchmark                                      Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------
fast_minify_twitter/repeats:10_mean        69493 ns        69491 ns           10 Gigabytes=9.08769G/s docs=14.3903k/s
fast_minify_twitter/repeats:10_median      69474 ns        69471 ns           10 Gigabytes=9.09031G/s docs=14.3944k/s
fast_minify_twitter/repeats:10_stddev       80.0 ns         80.5 ns           10 Gigabytes=10.4996M/s docs=16.626/s
fast_minify_twitter/repeats:10_max         69712 ns        69712 ns           10 Gigabytes=9.09667G/s docs=14.4045k/s
fast_minify_gsoc/repeats:10_mean          370108 ns       370098 ns           10 Gigabytes=8.99177G/s docs=2.70199k/s
fast_minify_gsoc/repeats:10_median        370227 ns       370210 ns           10 Gigabytes=8.98904G/s docs=2.70117k/s
fast_minify_gsoc/repeats:10_stddev           470 ns          468 ns           10 Gigabytes=11.3825M/s docs=3.42041/s
fast_minify_gsoc/repeats:10_max           370571 ns       370568 ns           10 Gigabytes=9.01268G/s docs=2.70828k/s
```

Main branch (GCC 10):
```
------------------------------------------------------------------------------------------------
Benchmark                                      Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------
fast_minify_twitter/repeats:10_mean        96723 ns        96720 ns           10 Gigabytes=6.52931G/s docs=10.3391k/s
fast_minify_twitter/repeats:10_median      96632 ns        96630 ns           10 Gigabytes=6.53539G/s docs=10.3487k/s
fast_minify_twitter/repeats:10_stddev        179 ns          178 ns           10 Gigabytes=12.015M/s docs=19.0256/s
fast_minify_twitter/repeats:10_max         97084 ns        97081 ns           10 Gigabytes=6.5411G/s docs=10.3578k/s
fast_minify_gsoc/repeats:10_mean          515451 ns       515444 ns           10 Gigabytes=6.45625G/s docs=1.94008k/s
fast_minify_gsoc/repeats:10_median        515526 ns       515509 ns           10 Gigabytes=6.45542G/s docs=1.93983k/s
fast_minify_gsoc/repeats:10_stddev           545 ns          542 ns           10 Gigabytes=6.79287M/s docs=2.04123/s
fast_minify_gsoc/repeats:10_max           516098 ns       516079 ns           10 Gigabytes=6.46822G/s docs=1.94368k/s
```

(This seems to be neutral on `parse_gsoc` and `ondemand`, probably because they don't call `compress`.)