@lemire 2021-08-30
It looks reasonable to me. Let us run the CI tests.