@lemire 2021-08-30
It looks like it passes our tests. It looks good and I plan to merge, but I need to do more review first.