@dougallj 2021-09-01
No rush - happy to make changes, fix issues or answer questions whenever :)

(Looking at some of the Cortex microarchitectures I'm increasingly convinced this PR should be an improvement there too, although I haven't been able to test. e.g. on Cortex A75, somewhat surprisingly, `tbl.8b` seems to have three times better throughput than `tbl.16b`? So that Apple-M1-neutral change would be a win. We're also about 25 instructions shorter overall, which shouldn't hurt. If anyone wants to get some idea of the generated code, you can try https://godbolt.org/z/71vPdd8a3 - that code doesn't exactly match the pull request, or the original, but the logic should be the same.)