@lemire 2021-09-01
I have reviewed it and it looks very good. The new code is simpler and is likely to be make better use of the ARM NEON ISA generally. Simpler code that matches more closely the ISA is often faster. The submitter has compelling test results. Fantastic work.